+++ src/crypto/hash.h
@@ -69,7 +69,7 @@ namespace Crypto {
     hash_extra_blake, hash_extra_groestl, hash_extra_jh, hash_extra_skein
   };
 
-  inline void blimp_hash(const unsigned char* input, Hash &output, int length, const unsigned char* salt, int salt_length) {
+  inline void blimp_hash(const unsigned char* input, Hash &output, uint64_t length, const unsigned char* salt, uint64_t salt_length) {
     balloon_blake(input, reinterpret_cast<char *>(&output), length, salt, salt_length);
   }
